code/requirements.txt
groq
rank_bm25
python-dotenv
sentence-transformers
scikit-learn
textblob
fastapi
uvicorn
